29/03  Actually works, Reliable at 150cm

+++ lib/IRsensor/IRsensor.h
@@ -10,7 +10,7 @@ class IRsensor{
         IRsensor(int IRpin);
         float readIR();
         float AreadIR();
-        float formulaRead();
+        double formulaRead();
 
     private:
         int _IRpin;
